Update node-operator.md
I made improvements aimed at enhancing the user experience of the **tell the Protocol Contract to freeze and stake the funds** section.

I inserted the minimum amount of QSP needed to stake to reduce the effort required by the user. I also made efforts to make it obvious to the user that we are using the wei denomination.