Update node-operator.md
When Karina was trying to operate a node, she did not understand how to properly insert her `[MyEtherWallet keystore file]` into `resources/keystore/default.json`. 

My edits are meant to increase the clarity in this step of the process. 

I also include a image in this PR, but it will not appear until the image I attempted to upload is approved.